% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{alder_check}
\alias{alder_check}
\title{Check an alder notebook and print diagnostics.}
\usage{
alder_check(path)
}
\arguments{
\item{path}{Path to an alder notebook.}
}
\value{
A data frame of diagnostics, invisibly.
}
\description{
Check an alder notebook and print diagnostics.
}
\examples{
\dontrun{alder_check("analysis.R")}
}
