% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{sql}
\alias{sql}
\title{Execute SQL against an explicit DBI connection or notebook data frames.}
\usage{
sql(query, conn = NULL)
}
\arguments{
\item{query}{A single SQL query string.}

\item{conn}{An optional DBI connection.}
}
\description{
Execute SQL against an explicit DBI connection or notebook data frames.
}
