% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alder-package.R
\docType{package}
\name{alder-package}
\alias{alder}
\alias{alder-package}
\title{Alder package}
\description{
Reactive notebooks for R.
}
\author{
\strong{Maintainer}: Carl Stone \email{carl@alder.dev}

Authors:
\itemize{
  \item Carl Stone \email{carl@alder.dev}
}

}
\keyword{internal}
