% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{alder_source}
\alias{alder_source}
\title{Evaluate an alder notebook in topological order.}
\usage{
alder_source(path, env = parent.frame())
}
\arguments{
\item{path}{Path to an alder notebook.}

\item{env}{Environment receiving definitions.}
}
\value{
The evaluation environment, invisibly.
}
\description{
Evaluate an alder notebook in topological order.
}
