% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{alder_test}
\alias{alder_test}
\title{Run cells marked \verb{#| test: true} with testthat.}
\usage{
alder_test(path)
}
\arguments{
\item{path}{Path to an alder notebook.}
}
\value{
testthat result records, invisibly.
}
\description{
Run cells marked \verb{#| test: true} with testthat.
}
